git-svn-id: https://svn.kuali.org/repos/kfs/branches/unlabeled-1.2.2@4570 05dbba50-812c-11dd-b2f5-a7aa3c7f87f7

+++ work/src/org/kuali/kfs/module/ld/businessobject/CalculatedSalaryFoundationTracker.java
@@ -1,30 +1,36 @@
 /*
- * Copyright 2005-2006 The Kuali Foundation.
+ * Copyright (c) 2004, 2005 The National Association of College and University 
+ * Business Officers, Cornell University, Trustees of Indiana University, 
+ * Michigan State University Board of Trustees, Trustees of San Joaquin Delta 
+ * College, University of Hawai'i, The Arizona Board of Regents on behalf of the 
+ * University of Arizona, and the r*smart group.
  * 
- * $Source: /opt/cvs/kfs/work/src/org/kuali/kfs/module/ld/businessobject/CalculatedSalaryFoundationTracker.java,v $
+ * Licensed under the Educational Community License Version 1.0 (the "License"); 
+ * By obtaining, using and/or copying this Original Work, you agree that you 
+ * have read, understand, and will comply with the terms and conditions of the 
+ * Educational Community License.
  * 
- * Licensed under the Educational Community License, Version 1.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * You may obtain a copy of the License at:
  * 
- * http://www.opensource.org/licenses/ecl1.php
+ * http://kualiproject.org/license.html
  * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,  DAMAGES OR OTHER 
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN 
+ * THE SOFTWARE.
  */
 
 package org.kuali.module.labor.bo;
 
-import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.LinkedHashMap;
 
 import org.kuali.core.bo.BusinessObjectBase;
-import org.kuali.core.bo.Options;
+import org.kuali.core.bo.user.Options;
 import org.kuali.core.util.KualiDecimal;
 import org.kuali.module.chart.bo.Account;
 import org.kuali.module.chart.bo.Chart;
@@ -33,7 +39,7 @@ import org.kuali.module.chart.bo.SubAccount;
 import org.kuali.module.chart.bo.SubObjCd;
 
 /**
- * 
+ * @author Kuali Nervous System Team (kualidev@oncourse.iu.edu)
  */
 public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
 
@@ -48,8 +54,8 @@ public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
     private Timestamp csfCreateTimestamp;
     private String csfDeleteCode;
     private KualiDecimal csfAmount;
-    private BigDecimal csfFullTimeEmploymentQuantity;
-    private BigDecimal csfTimePercent;
+    private Integer csfFullTimeEmploymentQuantity;
+    private Integer csfTimePercent;
     private String csfFundingStatusCode;
     private Integer employeeRecord;
     private String earnCode;
@@ -308,7 +314,7 @@ public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
      * @return - Returns the csfFullTimeEmploymentQuantity
      * 
      */
-    public BigDecimal getCsfFullTimeEmploymentQuantity() {
+    public Integer getCsfFullTimeEmploymentQuantity() {
         return csfFullTimeEmploymentQuantity;
     }
 
@@ -318,7 +324,7 @@ public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
      * @param - csfFullTimeEmploymentQuantity The csfFullTimeEmploymentQuantity to set.
      * 
      */
-    public void setCsfFullTimeEmploymentQuantity(BigDecimal csfFullTimeEmploymentQuantity) {
+    public void setCsfFullTimeEmploymentQuantity(Integer csfFullTimeEmploymentQuantity) {
         this.csfFullTimeEmploymentQuantity = csfFullTimeEmploymentQuantity;
     }
 
@@ -329,7 +335,7 @@ public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
      * @return - Returns the csfTimePercent
      * 
      */
-    public BigDecimal getCsfTimePercent() {
+    public Integer getCsfTimePercent() {
         return csfTimePercent;
     }
 
@@ -339,7 +345,7 @@ public class CalculatedSalaryFoundationTracker extends BusinessObjectBase {
      * @param - csfTimePercent The csfTimePercent to set.
      * 
      */
-    public void setCsfTimePercent(BigDecimal csfTimePercent) {
+    public void setCsfTimePercent(Integer csfTimePercent) {
         this.csfTimePercent = csfTimePercent;
     }
